refactor: move mmu and cartridge modules to their own subdirectories and update paths in the codebase
