+++ tests/cpu/instructions/alu/add_sub.cpp
@@ -15,7 +15,7 @@
 #include <gtest/gtest.h>
 
 // boyboy
-#include "boyboy/mmu_constants.h"
+#include "boyboy/mmu/constants.h"
 
 // Helpers
 #include "helpers/cpu_fixtures.h"
