+++ tests/cpu/instructions/load_store/ld_r_hl.cpp
@@ -10,7 +10,7 @@
 #include <gtest/gtest.h>
 
 // boyboy
-#include "boyboy/mmu_constants.h"
+#include "boyboy/mmu/constants.h"
 
 // Helpers
 #include "helpers/cpu_fixtures.h"
