+++ tests/cpu/test_interrupts.cpp
@@ -12,7 +12,7 @@
 #include "boyboy/cpu/cpu_constants.h"
 #include "boyboy/cpu/interrupt_handler.h"
 #include "boyboy/cpu/interrupts.h"
-#include "boyboy/mmu_constants.h"
+#include "boyboy/mmu/constants.h"
 
 // Helpers
 #include "helpers/cpu_fixtures.h"
